大功告成

+++ app/src/main/java/com/example/travelingagent/myclass/Sight.java
@@ -4,11 +4,11 @@ import java.lang.*;
 
 
 public class Sight extends Spot{
-	protected int popularity;
-	protected int score, environment, service;
-	protected int menPiao;
+	protected double popularity;
+	protected double score, environment, service;
+	protected double menPiao;
 	
-	public Sight(String name, int ID, int spotType, String description, double longitude, double latitude, int popularity, int score, int environment, int service, int menPiao) {
+	public Sight(String name, int ID, int spotType, String description, double longitude, double latitude, double popularity, double score, double environment, double service, double menPiao) {
 		super(name, ID, spotType, description, longitude, latitude);
 		this.popularity = popularity;
 		this.score = score;
